requirements.txt
python-multipart
fastapi==0.95.2
uvicorn==0.22.0
tensorflow==2.12.0
tensorflow-hub
numpy==1.23.5
image==1.5.33
Pillow==9.4.0